requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==4.4.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
numpy==2.2.2
Pillow==9.5.0
proglog==0.1.10
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
